mlp -- added the concept of a "rcdaq name"
We are ften running more than one instance of rcdaq. For example,
at FermiLab, we had the main DAQ and the "slave DAQ" reading out
the wirechambers. In addiion, we had a number of other rcdaq instances
running for "parasitic" experiments.

In order to reduce the potential or confusion, each rcdaq instance
has now a name. It defaults to the hostname, but can be set to anything,
for example

daq_setname "MWPC Readout RCDAQ"

It will be present in the GUIs and can be obtained with daq_getname.