The web controls now start on the start of the server on the default port (8899).
I changed the web sockets so that the browser javascript is aware of the run and open state,
and sends the proper daq_begin/daq_end and daq_open/daq_close command. Before the
command would toggle the state. This makes it now possible to use a curl or wget command
to trigger a specific action.