mlp - added "immediate daq_end" for the benfit of bulk operations.
A normal daq_end ends synchronously. It waits until the last buffer is written out,
and after it returns, rcdaq is ready to go. When we are dealing with an array of RCDAQs,
the asynchronous end is a lot faster. You just issue the command, and the call returns
immediately.

This is called in rcdaq_client with daq_end -i.