mlp - changed the protocol with the remote server.
Before we owuld open the server once and leave the connection open. We would then take many runs and later close. This caused iussues if there was a long gap (hours) between runs that the network in between might close the socket. Now we open the socket, start a run, end, and close the socket.  It's a minor overhead for increased operational stability.

The other change is that rcdaq_server now maintains a state that can bet set whether or not to log locally to a file or to a server. daq_open is now takeing the proper action. This allows the GUIs to function better, as the state can be set at initialization.

I also cleaned up the daq_status output a bit and adapted the GUIs (including the websocket GUI) to the new "we are logging to a server" state.